Documents/Everis(NTT Data)/DA Lab J2J/demo-lab/src/main/webapp/app/entities/type/update/type-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { TypeService } from '../service/type.service';
import { IType, Type } from '../type.model';

import { TypeUpdateComponent } from './type-update.component';

describe('Type Management Update Component', () => {
  let comp: TypeUpdateComponent;
  let fixture: ComponentFixture<TypeUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let typeService: TypeService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [TypeUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(TypeUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(TypeUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    typeService = TestBed.inject(TypeService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should update editForm', () => {
      const type: IType = { id: 456 };

      activatedRoute.data = of({ type });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(type));
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Type>>();
      const type = { id: 123 };
      jest.spyOn(typeService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ type });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: type }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(typeService.update).toHaveBeenCalledWith(type);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Type>>();
      const type = new Type();
      jest.spyOn(typeService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ type });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: type }));
      saveSubject.complete();

      // THEN
      expect(typeService.create).toHaveBeenCalledWith(type);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Type>>();
      const type = { id: 123 };
      jest.spyOn(typeService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ type });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(typeService.update).toHaveBeenCalledWith(type);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });
});
